added environment helper, moved the tests to separate modules, multiple updates
